sizeof: consolidate hardware size constants
With CL 241679, there exists a safe, portable, and compile time
constant sizeof package. For consisency, all symbols dealing with
hardware size information are derived from it and tests have been left
in place to verify correctness.

One exception is cmd/compile/internal/ssa/poset.go, which because of
its use in the compiler bootstrapping process, was not modified.
Furthermore, because math/bits is a core package, it was also necessary
to ammend cmd/go's mock goroot and go/build's dependency tree testing.

Updates #29982

Change-Id: I0057ae5d3ce0accb1cbfbc1bdbece7f048dd4be3