cmd/compile: optimize unsigned comparisons to 0/1
The rules in #21439 can be split into two groups.

1. Since an unsigned integer >= 0 is always true and < 0 is always
false, it is architecture-independent. This CL adds these optimizations
to generic rules.

2. For other rules in #21439, we can save a comparison instruction that
followed by a conditional branch on an architecture that supports
compare-with-zero-and-branch instructions. For example,

  if x > 0 { ... } else { ... }

the original version:
  CMP $0, R0
  BLS 9

the optimized version:
  CBZ R0, 8

This CL adds the second group to only arm64, other architectures can
follow this patch.

Updates #21439

Change-Id: Iec7e3040b761ecb1e60908f764815fdd9bc62495