cmd/compile:  can't call SetABI() on a LSYM after it has been added to hash table
Fix bug in the code dealing with the alias/wrapper needed for conversion from
ABIInternal to ABI0 (needed for assembly language functions). We are doing a
LookupABIInit() (via nam.Sym.Linksym()) to enter the base symbol (f.lsym) into the
ABI hash tables, but then we are doing f.lsym.SetABI(obj.ABI0), which is a bug,
since we are not changing which hash table (ctxt.has vs ctxt.funchash) the symbol
is stored in.  We now have an ABI0 symbol in the wrong hash table.

DO NOT SUBMIT

Change-Id: Ifda5f697600a7f1fedda2f7a3fb3ab86f67ae154