runtime: validate candidate searchAddr in pageAlloc.alloc
Currently pageAlloc.find attempts to find a better estimate for the
first free page in the heap, even if the space its looking for isn't
necessarily going to be the first free page in the heap (e.g. if npages
>= 2). However, in doing so it has the potential to return a searchAddr
candidate that doesn't actually correspond to mapped memory, but this
candidate might still be adopted. As a result, pageAlloc.alloc's fast
path may look at unmapped summary memory and segfault. This case is rare
on most operating systems since the heap is kept fairly contiguous, so
the chance that the candidate searchAddr discovered is unmapped is
fairly low. Even so, this is totally possible and outside the user's
control when it happens (in fact, it's likely to happen consistently for
a given user on a given system).

Fix this problem by ensuring that our candidate always points to mapped
memory. We do this by looking at mheap's arenas structure first. If it
turns out our candidate doesn't correspond to mapped memory, then we
look at inUse to round up the searchAddr to the next mapped address.

Also, document that find may return a candidate searchAddr that doesn't
refer to mapped memory, and so it's the caller's responsibility to
verify. Fixing it this way turns out to be simpler because
pageAlloc.find has exactly one caller, but multiple exit points.

Fixes #40191.

Change-Id: I759efec78987e4a8fde466ae45aabbaa3d9d4214