cmd/go: annotate ImportPathsQuiet and modload.(*loader).doPkg
Since doPkg is run using a par.Work, context is also threaded
through par.Work, but I expect that code to be temporary because
bcmills's lazy loading work should eliminate the par.Work. This
is the only remaining usage of par.Work, so once that's in
par.Work will be gone.

Updates #38714

Change-Id: I63ad264e69da555c58838807ab53111a85ad9638