database/sql: add on-connect hook support
Hi, I know this PR will *not* be merged, but I wanted to propose some kind of solution giving users an ability to execute some queries every time a new connection is used.

Here is why I need this: I want to set a session-level variable, DB doesn't matter (I've tried on MySQL/MariaDB). I've googled a lot and wasn't able to find any way to do that correctly. `database/sql` maintains its own pool of connections and it is not enough to set this once on DB connect, I also don't want to prepend this to every SQL call, I just want to have a way to declare some kind of callback that happens every time when `database/sql` is creating a new connection (and not reusing the existing one).

This is quite close to what I want (just need to define a handler for `state=2`) but I can't use `DB` object there because DB != connection, and I cannot make `driverConn` available to the outside, so I can only pass a list of queries and their args to be executed in an event of new connection create.

I'm new to golang code, so this PR will be rejected 100%, but can you point me in the right direction, please?

Change-Id: I1f68e93fd3174968491565e504bccac839d96cfd
GitHub-Last-Rev: 3c534f6217ed687c3b45ada9162aa3c86abb6c5d
GitHub-Pull-Request: golang/go#40384