runtime: fixes so fp is correct for open-coded defers called in same frame after a recovery
When we call the deferreturn stub, in order to run any remaining defers after a
recovery, we need to restore the fp properly (as well as setting the pc and sp).
In order to do this, I needed to add a new field 'sigcode2' to the g struct, and
also a new help routine getcallerfp(). The implementation for the architectures
besides amd64 and arm64 came from Keith.

Also, this includes fixes from Keith to restore the fp properly for non-open-coded
defer from Keith.

Change-Id: Idfdb9541d3e4b646c12d435e43083c71b913b483