Fixes so that fp is correct for open-coded defers called in same frame after a recovery
When we call the deferreturn stub, in order to run any remaining defers after a
recovery, we need restore the fp properly (as well as setting the pc and sp).

Includes test code from Keith (test.go, PrintStack) and definition of getcallerfp,
that will be removed.

Haven't tried to figure out if we can write test code for this yet.

DO NOT SUBMIT

Change-Id: Idfdb9541d3e4b646c12d435e43083c71b913b483