runtime: fixes so fp is correct for open-coded defers called in same frame after a recovery
When we call the deferreturn stub, in order to run any remaining defers after a
recovery, we need to restore the fp properly (as well as setting the pc and sp).
In order to do this, I needed to add a new field 'sigcode2' to the g struct, and
also a new help routine getcallerfp() (currently only defined for amd64 and
arm64).

Includes test code from Keith (test.go, PrintStack) that will be removed before checkin.

Change-Id: Idfdb9541d3e4b646c12d435e43083c71b913b483