cmd/compile: re-check escape info when initializing wrapper struct of closure
When generating wrapper struct for closure, we inherit escape info from
the closure itself. This can lead to wrong escaping decision. The
genearted struct has the form:

	clos = &struct{F uintptr; R T}{M.T·f, x}

for x.M with x has type T. Here, x.M can be stack allocated, but the
clos can not. We need to double check whether clos must be heap
allocated or not and set the escape class correctly.

Fixes #39292

Change-Id: I14c1e591681d9d92317bb2396d6cf5207aa93e08