cmd/compile: re-check escape info when initializing wrapper struct of closure
When generating wrapper struct for closure, we inherit escape info from
the closure itself. This can lead to wrong escaping decision. The
genearted struct has the form:

	clos = &struct{F uintptr; R T}{T.M·f, x}

for x.M with x has type T. Here, x.M can be stack allocated, but x may
be not, cause the wrapper struct heap allocated, too. When initializing
the struct, we need to double check whether the struct must be heap
allocated or not and set the escape class correctly, instead of
inheriting from "x.M".

Fixes #39292

Change-Id: I14c1e591681d9d92317bb2396d6cf5207aa93e08