cmd/compile: move expandCalls earlier to re-enable move optimizations
late opt replaces
  MOVE A -> B, MOVE B -> C
with
  MOVE A -> B, MOVE A -> C
to create an opportunity for dead auto elim to completely remove b.

However these need to be exposed as moves -- the optimization ensures
that there is no overlap so it is safe, to see the overlap, it needs to
see moves to and from stack addresses.  In order for this to happen
when the moves are part of either result copying (after a call) or
argument copying (before a call), calls have to be expanded to do
whatever memory operations they need to do.

Change-Id: I8a70f6bebbc01c455aea8eadad0fd0858368c1e3