[dev.go2go] types2: update to include the latest go/types changes
This brings types2 (the syntax-based type-checker) up-to-date
with the latest version of the prototype go/types version. It
does not yet include the *B.go2 tests (using square brackets)
or the relevant syntax parser changes.

Test that check error positions continue to be disabled since
the synax parser records slightly different positions that the
go/parser (to be fixed).

Verified that selected *.go2 tests that don't have errors pass.
Type-checks the std library without errors.

Change-Id: If859b180bba8362db86c211eeb4dc2d46255df36