runtime: move malloc stats into consistentHeapStats
This change moves the mcache-local malloc stats into the
consistentHeapStats structure so the malloc stats can be managed
consistently with the memory stats.

The one risk with this change is the addition of a few operations on the
tiny allocation fast path. In practice, I believe this performance
impact will be small (TODO: benchmark it) but is likely to show a
regression in microbenchmarks. Tiny allocations are not critical to
memory statistics, so perhaps it would be OK to let them go out-of-sync
with the rest.

Change-Id: I2a4b75f209c0e659b9c0db081a3287bf227c10ca