cmd/compile: enable late expansion for interface calls
Includes a few tweaks to Value.copyOf(a) (make it a no-op for
a self-copy) and new pattern hack "___" is like ellipsis where
the other side doesn't need to have matching ellipsis.

Moved the arg-length check BEFORE the args are probed,
because not all instances of variable length OpFoo
will have all the args mentioned in some rule for OpFoo.

Change-Id: I66de40672b3794a6427890ff96c805a488d783f4