cmd/compile: fix live variable computation for deferreturn
Taking the live variable set from the last return point is problematic.
See #40629 for details, but there may not be a return point, or it may
be before the final defer.

Additionally, keeping track of the last call as a *Value doesn't quite
work, as if it is dead-code eliminated, its live variable information,
if it is available at all, is taken from a random other instruction.

Instead, just mark all the open-defer argument slots as live
throughout the function, and zero initialize them at the start of the
function.

Q: how does this CL interact with go115ReduceLiveness?

Fixes #40629

Change-Id: Ie456c7db3082d0de57eaa5234a0f32525a1cce13